requirements.txt
torch==2.2.2
torchvision==0.17.2
pennylane==0.36.0
autoray==0.6.9
pillow==10.3.0
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
captum==0.7.0
